chore: switch to react navigation
