+++ example/package.json
@@ -6,14 +6,14 @@
   "private": true,
   "main": "main.js",
   "dependencies": {
-    "@expo/ex-navigation": "^2.0.0",
     "@expo/vector-icons": "^4.0.0",
     "color": "^0.11.4",
     "escape-string-regexp": "^1.0.5",
     "expo": "~16.0.0",
     "react": "16.0.0-alpha.6",
     "react-native": "git+https://github.com/expo/react-native.git#sdk-16.0.0",
-    "react-native-drawer": "^2.3.0"
+    "react-native-drawer": "^2.3.0",
+    "react-navigation": "^1.0.0-beta.9"
   },
   "devDependencies": {
     "babel-plugin-module-resolver": "^2.7.0",
